Added fulmar/typed.
Specify #lang fulmar/type to write typed code. Unfortunately I'm pretty
sure there's a contract check as each top level form is passed to
write-chunk, but I'm as of yet unsure how to prove that or remove it if
present. fulmar/typed also doesn't support the define/doc documentation
form, but I think the purpose of that is partially satisfied by type
declarations so we'll need to reconsider that issue anyway.